server/requirements.txt
# python==3.11
celery==5.3.6
django==4.2.8
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==23.5
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psutil==5.9.6

redis==5.0.4
pymysql==1.1.0
pymongo==4.7.2
